fix adventure save/load in TUI
