+++ osrgame/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "osrgame"
-version = "0.1.2"
+version = "0.1.3"
 description = "Text user interface (TUI) for OSR Console, a turn-based dungeon crawler RPG in the old-school renaissance style."
 authors = ["Marsh Macy <mmacy@users.noreply.github.com>"]
 license = "MIT"
@@ -10,7 +10,6 @@ readme = "README.md"
 python = "^3.11"
 textual = "^0.40.0"
 osrlib = { path = "../osrlib" }
-textual-dev = "^1.2.1"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.9.1"
@@ -18,6 +17,7 @@ flake8 = "^6.1.0"
 flake8-docstrings = "^1.7.0"
 pytest = "^7.4.2"
 mkdocs-material = "^9.4.6"
+textual-dev = "^1.2.1"
 
 [build-system]
 requires = ["poetry-core"]
