+++ osrlib/osrlib/enums.py
@@ -12,6 +12,16 @@ class CharacterClassType(Enum):
     THIEF = "Thief"
     COMMONER = "Commoner"
 
+
+class OpenAIModelVersion(Enum):
+    """Specifies the version of the OpenAI model to use."""
+    DEFAULT = "gpt-3.5-turbo-1106"
+    GPT35 = "gpt-3.5"
+    GPT35TURBO = "gpt-3.5-turbo-1106"
+    GPT4 = "gpt-4"
+    GPT4TURBO = "gpt-4-1106-preview"
+
+
 class PartyReaction(Enum):
     """Specifies the reaction of a PC or monster party at the start of an encounter."""
