+++ osrlib/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "osrlib"
-version = "0.1.9"
+version = "0.1.12"
 description = "Turn-based dungeon-crawler game engine for OSR-style RPGs."
 authors = ["Marsh Macy <mmacy@users.noreply.github.com>"]
 license = "MIT"
@@ -10,17 +10,17 @@ readme = "README.md"
 python = "^3.11"
 tinydb = "^4.8.0"
 openai = "^1.3.3"
-python-dotenv = "^1.0.0"
-pytest-mock = "^3.12.0"
-pytest-rerunfailures = "^12.0"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.9.1"
 flake8 = "^6.1.0"
 flake8-docstrings = "^1.7.0"
-pytest = "^7.4.2"
 mkdocs-material = "^9.4.6"
+pytest = "^7.4.2"
 pytest-cov = "^4.1.0"
+pytest-mock = "^3.12.0"
+pytest-rerunfailures = "^12.0"
+python-dotenv = "^1.0.0"
 
 [build-system]
 requires = ["poetry-core"]
