+++ tests/pytest.ini
@@ -1,7 +1,7 @@
 # pytest.ini
 [pytest]
-log_cli = true
-log_cli_level = INFO
+log_cli = false
+log_cli_level = DEBUG
 log_cli_format = %(asctime)s [%(levelname)s][%(module)s::%(funcName)s] %(message)s
 log_cli_date_format = %Y-%m-%d %H:%M:%S
 markers =
