+++ tests/test_unit_dice_roller.py
@@ -6,7 +6,7 @@ import pytest
 from osrlib.dice_roller import DiceRoll, roll_dice
 
 logger = logging.getLogger(__name__)
-logger.info("Testing dice_roller.py")
+logger.debug("Testing dice_roller.py")
 
 
 def test_valid_notation():
